Returned to urls from paths
