+++ news/serializer.py
@@ -1,16 +1,7 @@
 from rest_framework import serializers
-from rest_framework.response import Response
-from rest_framework.views import APIView
-from .models import  MoringaMerch
-from .serializer import MerchSerializer
+from .models import MoringaMerch
 
 class MerchSerializer(serializers.ModelSerializer):
     class Meta:
         model = MoringaMerch
         fields = ('name', 'description', 'price')
-
-class MerchList(APIView):
-    def get(self, request, format=None):
-        all_merch = MoringaMerch.objects.all()
-        serializers = MerchSerializer(all_merch, many=True)
-        return Response(serializers.data)
