+++ news/urls.py
@@ -1,15 +1,17 @@
-from django.urls import path
+
+  
+from django.conf.urls import url
 from . import views
 from django.conf import settings
 from django.conf.urls.static import static
 
 urlpatterns=[
-    path('^$',views.news_today,name='newsToday'),
-    path('^archives/(\d{4}-\d{2}-\d{2})/$',views.past_days_news,name = 'pastNews'),
-    path('^search/', views.search_results, name='search_results'),
-    path('^article/(\d+)',views.article,name ='article'),
-    path('^new/article$', views.new_article, name='new-article'),
-    path('^ajax/newsletter/$', views.newsletter, name='newsletter')
+    url(r'^$',views.news_today,name='newsToday'),
+    url(r'^archives/(\d{4}-\d{2}-\d{2})/$',views.past_days_news,name = 'pastNews'),
+    url(r'^search/', views.search_results, name='search_results'),
+    url(r'^article/(\d+)',views.article,name ='article'),
+    url(r'^new/article$', views.new_article, name='new-article'),
+    url(r'^api/merch/$', views.MerchList.as_view())
 
 ]
 if settings.DEBUG:
