+++ news/views.py
@@ -8,6 +8,11 @@ from . emails import send_welcome_email
 from django.contrib.auth.decorators import login_required
 from django.http import JsonResponse
 
+from rest_framework.response import Response
+from rest_framework.views import APIView
+from .models import  MoringaMerch
+from .serializer import MerchSerializer
+
 
 
 # View Function to present news from past days
@@ -111,3 +116,10 @@ def new_article(request):
     else:
         form = NewArticleForm()
     return render(request, 'new_article.html', {"form": form})
+
+
+class MerchList(APIView):
+    def get(self, request, format=None):
+        all_merch = MoringaMerch.objects.all()
+        serializers = MerchSerializer(all_merch, many=True)
+        return Response(serializers.data)
