+++ requirements.txt
@@ -1,18 +1,23 @@
-asgiref==3.4.1
-backports.zoneinfo==0.2.1
-config==0.5.1
+asgiref==3.2.10
+beautifulsoup4==4.9.1
+certifi==2020.6.20
+cloudinary==1.21.1
 confusable-homoglyphs==3.2.0
 dj-database-url==0.5.0
-Django==4.0
-django-bootstrap3==21.1
+Django==3.0.8
+django-bootstrap3==14.1.0
+django-bootstrap4==2.2.0
 django-heroku==0.3.1
-django-registration==3.2
-django-tinymce==3.4.0
-djangorestframework==3.12.4
-gunicorn==20.1.0
-Pillow==8.4.0
-psycopg2==2.9.2
-python-decouple==3.5
-pytz==2021.3
-sqlparse==0.4.2
-whitenoise==5.3.0
+django-registration==3.1
+djangorestframework==3.11.0
+gunicorn==20.0.4
+importlib-metadata==1.7.0
+psycopg2==2.8.5
+python-decouple==3.3
+pytz==2020.1
+six==1.15.0
+soupsieve==2.0.1
+sqlparse==0.3.1
+urllib3==1.25.9
+whitenoise==5.1.0
+zipp==3.1.0
